// ProjectMomoDonation.FrontEnd/src/app/feature/user/model/donate-response-query-params.model.ts
export interface DonationResponseQueryParams {
    partnerCode: string;
    orderId: string;
    requestId: string;
    amount: number;
    orderInfo: string;
    orderType: string;
    transId: number;
    resultCode: number;
    message: string;
    payType: string;
    responseTime: number;
    extraData: string;
    signature: string;
  }